Refactor the main page
