+++ frontend/src/app/signin/page.tsx
@@ -9,7 +9,7 @@ function Page() {
   const { user } = useAuthContext();
 
   if (!!user) {
-    router.push("/home")
+    router.push("/")
   }
 
   const [email, setEmail] = React.useState('')
@@ -33,7 +33,7 @@ function Page() {
     }
 
     console.log(result)
-    return router.push("/home")
+    return router.push("/")
   }
   return (
     <div className="d-flex justify-content-center align-items-center h-100 mt-5">
